move around more files
