+++ packages/core/src/connectivity/scp-cf/destination/destination-cache.ts
@@ -1,5 +1,6 @@
-import { tenantId, userId } from '../../../util';
 import { Cache, IsolationStrategy } from '../cache';
+import { tenantId } from '../tenant';
+import { userId } from '../user';
 import { Destination } from './destination-service-types';
 import { DestinationsByType } from './destination-accessor-types';
