+++ packages/core/src/connectivity/scp-cf/destination/destination-from-env.ts
@@ -3,7 +3,7 @@ import {
   proxyStrategy,
   ProxyStrategy,
   addProxyConfigurationInternet
-} from '../../../util';
+} from '../proxy-util';
 import {
   sanitizeDestination,
   isDestinationConfiguration,
