+++ packages/core/src/connectivity/scp-cf/index.ts
@@ -1,3 +1,4 @@
+export * from './jwt';
 export * from './cache';
 export * from './client-credentials-token';
 export * from './client-credentials-token-cache';
@@ -10,13 +11,14 @@ export * from './resilience-options';
 export * from './token-accessor';
 export * from './xsuaa-service';
 export * from './xsuaa-service-types';
-export * from './protocol'
-export * from './authorization-header'
-export * from './csrf-token-header'
-export * from './get-protocol'
-export * from './header-builder-for-destination'
+export * from './protocol';
+export * from './authorization-header';
+export * from './csrf-token-header';
+export * from './get-protocol';
+export * from './proxy-util';
+export * from './header-builder-for-destination';
 export {
-  mapping as mappingTenantFields,
+  mappingTenantFields,
   Tenant,
   tenantFromJwt,
   RegisteredJWTClaimsTenant
@@ -24,6 +26,6 @@ export {
 export {
   RegisteredJWTClaimsUser,
   Scope,
-  mapping as mappingUserFields,
+  mappingUserFields,
   userFromJwt
 } from './user';
