+++ packages/core/src/connectivity/scp-cf/tenant.ts
@@ -2,17 +2,40 @@ import {
   checkMandatoryValue,
   DecodedJWT,
   JwtKeyMapping,
-  tenantId,
-  tenantName
-} from '../../util';
+  readPropertyWithWarn
+} from './jwt';
 
 /**
  * Mapping between key name in the Tenant and key name in decoded JWT and the
  */
-export const mapping: JwtKeyMapping<Tenant, RegisteredJWTClaimsTenant> = {
+export const mappingTenantFields: JwtKeyMapping<
+  Tenant,
+  RegisteredJWTClaimsTenant
+> = {
   id: { keyInJwt: 'zid', extractorFunction: tenantId },
   name: { keyInJwt: 'zdn', extractorFunction: tenantName }
 };
+/**
+ * Get the tenant id of a decoded JWT.
+ * @param decodedToken - Token to read the tenant id from.
+ * @returns The tenant id if available.
+ */
+export function tenantId(decodedToken: DecodedJWT): string | undefined {
+  return readPropertyWithWarn(decodedToken, mappingTenantFields.id.keyInJwt);
+}
+
+/**
+ * Get the tenant name of a decoded JWT.
+ * @param decodedToken - Token to read the tenant id from.
+ * @returns The tenant id if available.
+ */
+export function tenantName(decodedToken: DecodedJWT): string | undefined {
+  const extAttr = readPropertyWithWarn(decodedToken, 'ext_attr');
+  if (extAttr) {
+    return readPropertyWithWarn(extAttr, 'zdn');
+  }
+  return undefined;
+}
 
 /**
  * Keys in the JWT related to the user
@@ -38,7 +61,7 @@ export interface Tenant {
  * @exception Error Raised if no id is found in the decoded JWT.
  */
 export function tenantFromJwt(decodedJWT: DecodedJWT): Tenant {
-  checkMandatoryValue('id', mapping, decodedJWT);
+  checkMandatoryValue('id', mappingTenantFields, decodedJWT);
   return {
     id: tenantId(decodedJWT)!,
     name: tenantName(decodedJWT)
