+++ packages/core/src/http-client/index.ts
@@ -1,2 +1,4 @@
 export * from './http-client';
 export * from './http-client-types';
+export * from './http-agent';
+export * from './agent-config';
