+++ packages/core/src/index.ts
@@ -1,10 +1,5 @@
-export * from './connectivity/scp-cf';
-export * from './util';
+export * from './connectivity';
 export * from './odata-common';
 export * from './odata-v2';
 export * from './odata-v4';
-export * from './http-agent';
-export * from './http-client';
-export * from './connectivity/scp-cf/protocol';
-export * from './connectivity/scp-cf/get-protocol';
 export * from './rest';
