+++ packages/core/src/odata-common/entity.ts
@@ -1,10 +1,11 @@
 /* eslint-disable max-classes-per-file */
 
 import { equal, isNullish } from '@sap-cloud-sdk/util';
-import { isNavigationProperty, nonEnumerable, toPropertyFormat } from '../util';
 import { EntityBuilder } from './entity-builder';
 import { Link, Field, Selectable, CustomFieldBase } from './selectable';
 import { RequestBuilder } from './request-builder';
+import { isNavigationProperty, nonEnumerable } from './properties-util';
+import { toPropertyFormat } from './name-converter';
 
 export type ODataVersionOf<T extends EntityBase> = T['_oDataVersion'];
