+++ packages/core/src/odata-common/request-builder/batch/batch-response-parser.ts
@@ -1,5 +1,10 @@
 /* Copyright (c) 2020 SAP SE or an SAP affiliate company. All rights reserved. */
-import { last, createLogger, errorWithCause, getHeaderValue } from '@sap-cloud-sdk/util';
+import {
+  last,
+  createLogger,
+  errorWithCause,
+  getHeaderValue
+} from '@sap-cloud-sdk/util';
 import { HttpResponse } from '../../../http-client';
 
 const logger = createLogger({
