+++ packages/core/src/odata-common/request/odata-request-config.ts
@@ -1,4 +1,8 @@
-import { createLogger, VALUE_IS_UNDEFINED, mergeHeaders } from '@sap-cloud-sdk/util';
+import {
+  createLogger,
+  VALUE_IS_UNDEFINED,
+  mergeHeaders
+} from '@sap-cloud-sdk/util';
 
 export type RequestMethodType = 'get' | 'post' | 'patch' | 'delete' | 'put';
