+++ packages/core/src/odata-common/request/odata-request.ts
@@ -1,19 +1,26 @@
-import { errorWithCause, propertyExists ,  filterNullishValues,
+import {
+  errorWithCause,
+  propertyExists,
+  filterNullishValues,
   getHeader,
   getHeaders,
   getHeaderValue,
-  mergeHeaders} from '@sap-cloud-sdk/util';
-import { Destination, sanitizeDestination } from '../../connectivity/scp-cf';
+  mergeHeaders
+} from '@sap-cloud-sdk/util';
+import {
+  Destination,
+  sanitizeDestination,
+  buildHeadersForDestination,
+  buildCsrfHeaders
+} from '../../connectivity/scp-cf';
 import {
   removeLeadingSlashes,
   removeSlashes,
   removeTrailingSlashes
-} from '../../util/remove-slashes';
+} from '../remove-slashes';
 import { HttpResponse, executeHttpRequest } from '../../http-client';
 import { ODataRequestConfig } from './odata-request-config';
 import { isWithETag } from './odata-request-traits';
-import { buildCsrfHeaders } from '../../connectivity/scp-cf';
-import { buildHeadersForDestination } from '../../connectivity/scp-cf';
 
 /**
  * OData request configuration for an entity type.
