--- packages/core/src/util/index.ts
@@ -1,5 +1,0 @@
-export * from './jwt';
-export * from './name-converter';
-export * from './properties-util';
-export * from './proxy-util';
-export * from './remove-slashes';
