+++ packages/core/test/request-builder/header-builder/authorization-header.spec.ts
@@ -8,7 +8,8 @@ import {
 } from '../../../src';
 import {
   addAuthorizationHeader,
-  buildAuthorizationHeaders
+  buildAuthorizationHeaders,
+  buildHeadersForDestination
 } from '../../../src/connectivity/scp-cf';
 import { ODataGetAllRequestConfig } from '../../../src/odata-common';
 import {
@@ -16,7 +17,6 @@ import {
   defaultBasicCredentials
 } from '../../test-util/request-mocker';
 import { TestEntity } from '../../test-util/test-services/v2/test-service';
-import { buildHeadersForDestination } from '../../../src/connectivity/scp-cf';
 
 describe('Authorization header builder', () => {
   it('does not throw on NoAuthentication', async () => {
