+++ packages/core/test/test-util/destination-service-mocks.ts
@@ -1,5 +1,5 @@
 import nock from 'nock';
-import * as sdkJwt from '../../src/util/jwt';
+import * as sdkJwt from '../../src/connectivity/scp-cf/jwt';
 import { destinationServiceUri } from './environment-mocks';
 
 type nockFunction = (a: string, b: nock.Options) => nock.Scope;
