+++ packages/core/test/test-util/token-accessor-mocks.ts
@@ -1,5 +1,5 @@
-import { decodeJwt } from '../../src/util';
 import * as tokenAccessor from '../../src/connectivity/scp-cf/token-accessor';
+import { decodeJwt } from '../../src/connectivity/scp-cf';
 import { TestTenants } from './environment-mocks';
 import {
   providerServiceToken,
